feat: enable cursor synching across all profiler charts
